primaImplementazioneNpcHandler
