+++ Unity/Yummy-verse/Assets/Prefabs/ominoconCamminata@Working On Device.fbx
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:9c19382dc9645e6eaaa8fde466a61acd480f9456bb0477d5c38256027fcf32e5
+size 3264288
