+++ Unity/Yummy-verse/Assets/Scripts/Movement/NpcBocca.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ed3a970a66efa0842a706a6ff4f93202
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
