+++ Unity/Yummy-verse/Assets/Scripts/Movement/NpcStomaco.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 51f9729a0949e0d49abaf6f5632c9b50
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
