+++ Unity/Yummy-verse/Assets/Scripts/Movement/NpcStomaco/NPC_Stomaco.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ae9d64ef335a8344baaf1fd452868a4b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
