+++ Unity/Yummy-verse/Assets/Scripts/Movement/NpcStomaco/Npc_stomaco.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 93c885aed61be4a4f9ef4d2bef765864
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
